warnings on win32 platform fixed
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@1100 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libzmq/zmq/pipe.hpp
@@ -123,7 +123,7 @@ namespace zmq
         bool delayed_gap;
 
         //  Number of messages kept in main memory.
-        size_t in_core_msg_cnt;
+        uint64_t in_core_msg_cnt;
 
 #ifdef ZMQ_HAVE_DATA_DAM
         //  Message store keeps messages when the memory buffer is full.
